Created switch class to instantiate physical OF switch
* The RyuREST API has been modified and split into 2 APIs:
1. ryu_switch: [Object Orientated Approach] Allows the user to
instantiate a switch object that will represent a single physical
switch connected to Ryu. REST API calls on this object affect the
defined switch itself.
2. ryu_rest: [Functional Approach] Class has been removed. This file is
not a module of functions that can be imported. These are the same
methods from ryu_switch, however the DPID is a parameter/argument used
to specify the target switch.